+++ tests/Upwork/API/Routers/Freelancers/SearchTest.php
@@ -10,7 +10,7 @@ class SearchTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }