+++ tests/Upwork/API/Routers/Hr/EngagementsTest.php
@@ -10,7 +10,7 @@ class EngagementsTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }