+++ tests/Upwork/API/Routers/Hr/RolesTest.php
@@ -10,7 +10,7 @@ class RolesTest extends CommonTestRouter
     /**
      * Setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }